get&set Nano_2.4_TX_2.0.1 config

+++ USER/Drivers/joystick.h
@@ -23,7 +23,7 @@ typedef enum
 }requestProtocol_e;
 
 
-extern uint8_t sendSpam;                                                                
+extern uint16_t sendSpam;                                                                
                                                                 
 void joystickTask(void *param);
 extern TaskHandle_t joystickTaskHandle;
